seperated input and alert component
