+++ src/screens/App/screens/Login/components/Login.jsx
@@ -2,46 +2,8 @@ import React, { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import axios from "../../../../../shared/caller";
 import { useForm } from "../../../../../shared/Form/useForm";
-
-function InputField({ type, name, label, error, value, onChange, svgD }) {
-  const errorValues = error ? "border-red-500" : "border-gray-100";
-
-  return (
-    <div className="flex flex-col space-y-1">
-      <div className="flex justify-between">
-        <label className="font-semibold text-md text-gray-500">{label}</label>
-        <label className="text-red-400 font-semibold ">{error}</label>
-      </div>
-
-      <div
-        className={`transition duration-300 inline-flex shadow-md rounded-md h-11 items-center px-3 space-x-2 bg-gray-100 border  ${errorValues}`}
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          className="h-6 w-6 text-my-accent"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d={svgD}
-          />
-        </svg>
-
-        <input
-          name={name}
-          type={type}
-          className="w-full focus:outline-none bg-gray-100"
-          value={value}
-          onChange={onChange}
-        />
-      </div>
-    </div>
-  );
-}
+import InputField from "../../../../../shared/Components/auth/InputField.Auth";
+import Alert from "../../../../../shared/Components/auth/Alert";
 
 function Login({ history }) {
   // state variable for API message
@@ -61,7 +23,6 @@ function Login({ history }) {
         let messageError;
         if (err.response) messageError = err.response.data.error;
         else messageError = "Something went wrong. Try again.";
-        console.log(messageError);
         setReqErr((prev) => messageError);
       });
   }
@@ -125,35 +86,7 @@ function Login({ history }) {
         </div>
 
         <div className="flex flex-col w-full gap-y-8">
-          <div
-            className={
-              (reqErr ? "absolute" : "hidden") +
-              " transition duration-300 h-14 w-72 bg-red-100 rounded-md shadow-md border-l-4 border-red-500 px-4 flex justify-between items-center  left-10 bottom-5"
-            }
-          >
-            <div className="text-red-800 font-semibold">{reqErr}</div>
-
-            <button
-              onClick={() => {
-                setReqErr("");
-              }}
-            >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-6 w-6"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            </button>
-          </div>
+          <Alert reqErr={reqErr} setReqErr={setReqErr} />
 
           <InputField
             name="email"
